+++ www-snippet/create-react-component/__mocks__/react-component/LoremIpsumWithRef_tsx
