+++ www-snippet/create-react-component/__mocks__/react-component/LoremIpsum_spec_tsx
