+++ www-snippet/create-react-component/__mocks__/react-component/LoremIpsum_stories_tsx
