+++ www-snippet/create-react-component/__mocks__/react-component/index_ts
