+++ www-snippet/create-react-component/__mocks__/readMockFile.test.ts
