+++ www-snippet/create-react-component/create-react-component-template.test.ts
@@ -5,7 +5,7 @@ import {
   createReactComponentTestingContent,
 } from "./create-react-component-template.ts";
 import { assertEquals } from "https://deno.land/std@0.106.0/testing/asserts.ts";
-import { readMockFile } from "./__mocks__/readMockFile.ts";
+import { readMockFile } from "../__mocks__/readMockFile.ts";
 
 Deno.test(
   `should return correct react component.`,
