+++ www-snippet/create-react-component/create-react-component-template.ts
